date-utils refactor
