+++ src/js/components/pages/greendash/CompareCard.jsx
@@ -9,9 +9,10 @@ import KStatus from '../../../base/data/KStatus';
 import { getDataList } from '../../../base/plumbing/Crud';
 import DataStore from '../../../base/plumbing/DataStore';
 import SearchQuery from '../../../base/searchquery';
-import { isoDate } from '../../../base/utils/miscutils';
 import C from '../../../C';
-import { dataColours, getPeriodQuarter, GreenCard, GreenCardAbout, ModeButton, printPeriod, TONNES_THRESHOLD } from './dashutils';
+import {GreenCard, GreenCardAbout, ModeButton} from './GreenDashUtils';
+import { dataColours, TONNES_THRESHOLD } from './dashUtils';
+import { isoDate, getPeriodQuarter, printPeriod } from '../../../base/utils/date-utils';
 import { getCompressedBreakdownWithCount, getCarbon, emissionsPerImpressions, getSumColumn } from './emissionscalcTs';
 
 import { isPer1000 } from './GreenMetrics';
