+++ src/js/components/pages/greendash/GreenDashboardFilters.jsx
@@ -5,7 +5,7 @@ import KStatus from '../../../base/data/KStatus';
 import DataStore from '../../../base/plumbing/DataStore';
 import { nonce } from '../../../base/data/DataClass';
 import { getDataItem, getDataList } from '../../../base/plumbing/Crud';
-import { getPeriodQuarter, getPeriodMonth, periodFromUrl, periodToParams, printPeriod } from './dashutils';
+import { getPeriodQuarter, getPeriodMonth, periodFromUrl, periodToParams, printPeriod } from '../../../base/utils/date-utils';
 
 import DateRangeWidget from '../../DateRangeWidget';
 import { modifyPage } from '../../../base/plumbing/glrouter';
